pyproject.toml
[tool.poetry]
name = "gridwatch-scilly-api"
version = "0.1.0"
description = "REST API for Gridwatch Scilly."
authors = ["George Goldberg <george@appletreetechnology.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
pytest = "^7.1.3"
requests = "^2.28.1"
SQLAlchemy = "^1.4.41"
psycopg2-binary = "^2.9.4"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
